@@ -12,12 +12,12 @@ class LoopBackSample extends StatefulWidget {
 }
 
 class _MyAppState extends State<LoopBackSample> {
-  MediaStream _localStream;
-  RTCPeerConnection _peerConnection;
+  MediaStream? _localStream;
+  RTCPeerConnection? _peerConnection;
   final _localRenderer = RTCVideoRenderer();
   final _remoteRenderer = RTCVideoRenderer();
   bool _inCalling = false;
-  Timer _timer;
+  Timer? _timer;
 
   String get sdpSemantics =>
       WebRTC.platformIsWindows ? 'plan-b' : 'unified-plan';
@@ -100,17 +100,13 @@ class _MyAppState extends State<LoopBackSample> {
   }
 
   void _onCandidate(RTCIceCandidate candidate) {
-    if (candidate == null) {
-      print('onCandidate: complete!');
-      return;
-    }
-    print('onCandidate: ' + candidate.candidate);
-    _peerConnection.addCandidate(candidate);
+    print('onCandidate: ${candidate.candidate}');
+    _peerConnection?.addCandidate(candidate);
   }
 
   void _onTrack(RTCTrackEvent event) {
     print('onTrack');
-    if (event.track.kind == 'video' && event.streams.isNotEmpty) {
+    if (event.track.kind == 'video' && (event.streams.isNotEmpty)) {
       print('New stream: ' + event.streams[0].id);
       _remoteRenderer.srcObject = event.streams[0];
     }
@@ -164,14 +160,14 @@ class _MyAppState extends State<LoopBackSample> {
       _peerConnection =
           await createPeerConnection(configuration, loopbackConstraints);
 
-      _peerConnection.onSignalingState = _onSignalingState;
-      _peerConnection.onIceGatheringState = _onIceGatheringState;
-      _peerConnection.onIceConnectionState = _onIceConnectionState;
-      _peerConnection.onConnectionState = _onPeerConnectionState;
-      _peerConnection.onAddStream = _onAddStream;
-      _peerConnection.onRemoveStream = _onRemoveStream;
-      _peerConnection.onIceCandidate = _onCandidate;
-      _peerConnection.onRenegotiationNeeded = _onRenegotiationNeeded;
+      _peerConnection!.onSignalingState = _onSignalingState;
+      _peerConnection!.onIceGatheringState = _onIceGatheringState;
+      _peerConnection!.onIceConnectionState = _onIceConnectionState;
+      _peerConnection!.onConnectionState = _onPeerConnectionState;
+      _peerConnection!.onAddStream = _onAddStream;
+      _peerConnection!.onRemoveStream = _onRemoveStream;
+      _peerConnection!.onIceCandidate = _onCandidate;
+      _peerConnection!.onRenegotiationNeeded = _onRenegotiationNeeded;
 
       _localStream =
           await navigator.mediaDevices.getUserMedia(mediaConstraints);
@@ -179,12 +175,12 @@ class _MyAppState extends State<LoopBackSample> {
 
       switch (sdpSemantics) {
         case 'plan-b':
-          await _peerConnection.addStream(_localStream);
+          await _peerConnection!.addStream(_localStream!);
           break;
         case 'unified-plan':
-          _peerConnection.onTrack = _onTrack;
-          _localStream.getTracks().forEach((track) {
-            _peerConnection.addTrack(track, _localStream);
+          _peerConnection!.onTrack = _onTrack;
+          _localStream!.getTracks().forEach((track) {
+            _peerConnection!.addTrack(track, _localStream!);
           });
           break;
       }
@@ -243,13 +239,13 @@ class _MyAppState extends State<LoopBackSample> {
           init:
               RTCRtpTransceiverInit(direction: TransceiverDirection.RecvOnly));
       */
-      var description = await _peerConnection.createOffer(offerSdpConstraints);
+      var description = await _peerConnection!.createOffer(offerSdpConstraints);
       var sdp = description.sdp;
       print('sdp = $sdp');
-      await _peerConnection.setLocalDescription(description);
+      await _peerConnection!.setLocalDescription(description);
       //change for loopback.
       description.type = 'answer';
-      await _peerConnection.setRemoteDescription(description);
+      await _peerConnection!.setRemoteDescription(description);
 
       /* Unfied-Plan replaceTrack
       var stream = await MediaDevices.getDisplayMedia(mediaConstraints);
@@ -271,8 +267,8 @@ class _MyAppState extends State<LoopBackSample> {
 
   void _hangUp() async {
     try {
-      await _localStream.dispose();
-      await _peerConnection.close();
+      await _localStream?.dispose();
+      await _peerConnection?.close();
       _peerConnection = null;
       _localRenderer.srcObject = null;
       _remoteRenderer.srcObject = null;
@@ -282,13 +278,13 @@ class _MyAppState extends State<LoopBackSample> {
     setState(() {
       _inCalling = false;
     });
-    _timer.cancel();
+    _timer?.cancel();
   }
 
   void _sendDtmf() async {
     var dtmfSender =
-        _peerConnection.createDtmfSender(_localStream.getAudioTracks()[0]);
-    await dtmfSender.insertDTMF('123#');
+        _peerConnection?.createDtmfSender(_localStream!.getAudioTracks()[0]);
+    await dtmfSender?.insertDTMF('123#');
   }
 
   @override
