@@ -7,8 +7,8 @@ import '../interface/rtc_track_event.dart';
 class RTCTrackEventWeb extends RTCTrackEvent {
   RTCTrackEventWeb(
       {RTCRtpReceiver? receiver,
-      List<MediaStream>? streams,
-      MediaStreamTrack? track,
+      required List<MediaStream> streams,
+      required MediaStreamTrack track,
       RTCRtpTransceiver? transceiver})
       : super(
             receiver: receiver,
